@@ -11,6 +11,7 @@
     "build:qa": "ng build -c=qa",
     "e2e": "ng e2e",
     "e2e:ci": "ng e2e --configuration=ci --protractor-config=projects/storefrontapp-e2e/protractor.headless.conf.js",
+    "generate:changelog": "ts-node ./scripts/changelog.ts",
     "lint": "ng lint",
     "mockserver": "cd ./projects/backend/mockgenerator && npm run start",
     "ng": "ng",
@@ -30,18 +31,18 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "~7.1.0",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/forms": "~7.1.0",
-    "@angular/http": "~7.1.0",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/platform-server": "^7.1.4",
-    "@angular/pwa": "^0.11.3",
-    "@angular/router": "~7.1.0",
-    "@angular/service-worker": "~7.1.0",
+    "@angular/animations": "~7.2.0",
+    "@angular/common": "~7.2.0",
+    "@angular/compiler": "~7.2.0",
+    "@angular/core": "~7.2.0",
+    "@angular/forms": "~7.2.0",
+    "@angular/http": "~7.2.0",
+    "@angular/platform-browser": "~7.2.0",
+    "@angular/platform-browser-dynamic": "~7.2.0",
+    "@angular/platform-server": "^7.2.0",
+    "@angular/pwa": "^0.12.0",
+    "@angular/router": "~7.2.0",
+    "@angular/service-worker": "~7.2.0",
     "@ng-bootstrap/ng-bootstrap": "^4.0.1",
     "@ng-select/ng-select": "^2.13.2",
     "@ngrx/effects": "~7.0.0",
@@ -52,25 +53,31 @@
     "@types/googlemaps": "^3.30.13",
     "bootstrap": "^4.1.3",
     "core-js": "^2.5.7",
-    "hamburgers": "^0.9.3",
+    "hamburgers": "^1.1.3",
     "ngrx-store-localstorage": "^5.1.0",
     "rxjs": "^6.3.3",
     "ts-loader": "^5.3.2",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.11.0",
-    "@angular-devkit/build-ng-packagr": "~0.11.0",
-    "@angular/cli": "~7.1.4",
-    "@angular/compiler-cli": "~7.1.0",
-    "@angular/language-service": "~7.1.0",
+    "@angular-devkit/build-angular": "~0.12.0",
+    "@angular-devkit/build-ng-packagr": "~0.12.0",
+    "@angular/cli": "~7.2.1",
+    "@angular/compiler-cli": "7.2.0",
+    "@angular/language-service": "~7.2.0",
     "@ngrx/store-devtools": "~7.0.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
+    "chalk": "^2.4.1",
     "codelyzer": "~4.5.0",
+    "commander": "^2.19.0",
     "concurrently": "^4.0.1",
+    "conventional-commits-parser": "^3.0.1",
+    "ejs": "^2.6.1",
     "faker": "^4.1.0",
+    "gh-got": "^8.0.1",
+    "git-raw-commits": "^2.0.0",
     "http-server": "^0.11.1",
     "jasmine-core": "~2.99.1",
     "jasmine-marbles": "^0.4.1",
@@ -88,6 +95,8 @@
     "node-sass": "^4.11.0",
     "prettier": "1.15.3",
     "protractor": "~5.4.0",
+    "semver": "^5.6.0",
+    "through2": "^3.0.0",
     "ts-node": "~7.0.0",
     "tsickle": "^0.34.0",
     "tslib": "^1.9.3",
