@@ -14,8 +14,8 @@
   },
   "repository": "https://github.com/SAP/cloud-commerce-spartacus-storefront",
   "peerDependencies": {
-    "@angular/common": "~7.1.0",
-    "@angular/core": "~7.1.0",
+    "@angular/common": "~7.2.0",
+    "@angular/core": "~7.2.0",
     "@ngrx/effects": "~7.0.0",
     "@ngrx/router-store": "~7.0.0",
     "@ngrx/store": "~7.0.0",
