@@ -1,22 +1,32 @@
 import {
   Component,
-  Input,
   ChangeDetectionStrategy,
-  OnDestroy,
-  ViewChild,
   OnInit,
+  ElementRef,
   ChangeDetectorRef
 } from '@angular/core';
 
-import { ProductService, Product, WindowRef } from '@spartacus/core';
+import {
+  ProductService,
+  Product,
+  CmsProductCarouselComponent,
+  WindowRef
+} from '@spartacus/core';
 
-import { Subscription, fromEvent, Observable } from 'rxjs';
-import { debounceTime } from 'rxjs/operators';
+import { fromEvent, Observable, of } from 'rxjs';
+import {
+  debounceTime,
+  map,
+  distinctUntilChanged,
+  startWith
+} from 'rxjs/operators';
 
-import { AbstractCmsComponent } from '../../cms/components/abstract-cms-component';
+import { CmsComponentData } from '../../cms/components/cms-component-data';
 
-import { CmsService } from '@spartacus/core';
-import { NgbCarousel } from '@ng-bootstrap/ng-bootstrap';
+const MAX_WIDTH = 360;
+const MAX_ITEM_SIZE = 4;
+
+const SPEED = 250;
 
 @Component({
   selector: 'cx-product-carousel',
@@ -24,97 +34,80 @@ import { NgbCarousel } from '@ng-bootstrap/ng-bootstrap';
   styleUrls: ['./product-carousel.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class ProductCarouselComponent extends AbstractCmsComponent
-  implements OnDestroy, OnInit {
-  productGroups: Array<string[]>;
-  products: { [key: string]: Observable<Product> } = {};
-
-  resize$: Subscription;
+export class ProductCarouselComponent implements OnInit {
+  items$: Observable<Observable<Product>[]>;
+  itemSize$: Observable<number>;
+  activeItem = 0;
 
   private window: Window;
 
-  @ViewChild('carousel')
-  carousel: NgbCarousel;
-
-  @Input()
-  productCodes: Array<string>;
-
   constructor(
-    protected cmsService: CmsService,
-    protected cd: ChangeDetectorRef,
+    public component: CmsComponentData<CmsProductCarouselComponent>,
     private productService: ProductService,
-    winRef: WindowRef
+    winRef: WindowRef,
+    private el: ElementRef,
+    private cd: ChangeDetectorRef
   ) {
-    super(cmsService, cd);
     this.window = winRef.nativeWindow;
   }
 
   ngOnInit() {
-    if (this.window) {
-      this.resize$ = fromEvent(this.window, 'resize')
-        .pipe(debounceTime(300))
-        .subscribe(() => this.createGroups());
+    this.setItemSize();
+    this.setItems();
+  }
+
+  /**
+   * Maps the item codes from CMS component to an array of `Product` observables.
+   */
+  protected setItems() {
+    this.items$ = this.component.data$.pipe(
+      map(data => {
+        const productCodes = data.productCodes.split(' ');
+        return productCodes.map(code => this.productService.get(code));
+      })
+    );
+  }
+
+  /**
+   * The number of items shown in the carousel can be calculated
+   * the standard implemenattions uses the element size to calculate
+   * the items that fit in the carousel.
+   * This method is called in `ngOnInit`.
+   */
+  protected setItemSize(): void {
+    if (!this.window) {
+      this.itemSize$ = of(MAX_ITEM_SIZE);
+      return;
     }
+    this.itemSize$ = fromEvent(this.window, 'resize').pipe(
+      map(() => (this.el.nativeElement as HTMLElement).clientWidth),
+      startWith((this.el.nativeElement as HTMLElement).clientWidth),
+      // avoid to much calls
+      debounceTime(100),
+      map((innerWidth: any) => {
+        const itemsPerPage = Math.round(innerWidth / MAX_WIDTH);
+        return itemsPerPage > 2 ? MAX_ITEM_SIZE : itemsPerPage;
+      }),
+      // only emit new size when the size changed
+      distinctUntilChanged()
+    );
   }
 
-  prev(): void {
-    this.carousel.prev();
+  prev(max) {
+    this.setActiveItem(this.activeItem - max, max);
   }
 
-  next(): void {
-    this.carousel.next();
+  next(max) {
+    this.setActiveItem(this.activeItem + max, max);
   }
 
-  protected fetchData(): void {
-    this.setProductCodes();
-    this.productCodes.forEach(code => {
-      this.products[code] = this.productService.get(code);
-    });
-    this.createGroups();
-    super.fetchData();
-  }
-
-  protected createGroups(): void {
-    const groups: Array<string[]> = [];
-    this.productCodes.forEach(product => {
-      const lastGroup: string[] = groups[groups.length - 1];
-      if (lastGroup && lastGroup.length < this.getItemsPerPage()) {
-        lastGroup.push(product);
-      } else {
-        groups.push([product]);
-      }
-    });
-    this.productGroups = groups;
-  }
-
-  protected getItemsPerPage(): number {
-    const smallScreenMaxWidth = 576;
-    const tabletScreenMaxWidth = 768;
-    const innerWidth = this.window ? window.innerWidth : tabletScreenMaxWidth;
-    let itemsPerPage: number;
-    if (innerWidth < smallScreenMaxWidth) {
-      itemsPerPage = 1;
-    } else if (
-      innerWidth > smallScreenMaxWidth &&
-      innerWidth < tabletScreenMaxWidth
-    ) {
-      itemsPerPage = 2;
-    } else {
-      itemsPerPage = 4;
-    }
-    return itemsPerPage;
-  }
-
-  protected setProductCodes(): void {
-    if (this.component && this.component.productCodes) {
-      this.productCodes = this.component.productCodes.split(' ');
-    }
-  }
-
-  ngOnDestroy() {
-    if (this.resize$) {
-      this.resize$.unsubscribe();
-    }
-    super.ngOnDestroy();
+  setActiveItem(newActive: number, max: number) {
+    this.activeItem = -1;
+    // we wait a little with setting the new active
+    // to make a better animation
+    setTimeout(() => {
+      this.activeItem = newActive;
+      this.cd.markForCheck();
+    }, (max - 1) * SPEED);
   }
 }
