@@ -1,16 +1,14 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { ActivatedRoute, Params } from '@angular/router';
-import { Store, select } from '@ngrx/store';
+import { ActivatedRoute } from '@angular/router';
 import { Observable, Subscription } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { tap } from 'rxjs/operators';
 
-import { SearchConfig } from '../../models/search-config';
-import { SearchQuery } from '../../models/search-query';
-import { LongitudeLatitude } from '../../models/longitude-latitude';
-import { StoreFinderService } from '../../services/store-finder.service';
-
-import * as fromStore from '../../store';
-
+import {
+  StoreFinderSearchConfig,
+  StoreFinderSearchQuery,
+  LongitudeLatitude,
+  StoreFinderService
+} from '@spartacus/core';
 @Component({
   selector: 'cx-store-finder-search-result',
   templateUrl: './store-finder-search-result.component.html',
@@ -18,58 +16,54 @@ import * as fromStore from '../../store';
 })
 export class StoreFinderSearchResultComponent implements OnInit, OnDestroy {
   locations: any;
-  searchQuery: SearchQuery;
+  searchQuery: StoreFinderSearchQuery;
   locations$: Observable<any>;
   isLoading$: Observable<any>;
   geolocation: LongitudeLatitude;
-  ngUnsubscribe: Subscription;
-  searchConfig: SearchConfig = {
+  subscription: Subscription;
+  searchConfig: StoreFinderSearchConfig = {
     currentPage: 0
   };
 
   constructor(
-    private store: Store<fromStore.StoresState>,
     private storeFinderService: StoreFinderService,
     private route: ActivatedRoute
   ) {}
 
-  ngOnInit(): void {
-    this.route.queryParams.subscribe(params => this.initialize(params));
+  ngOnInit() {
+    this.subscription = this.route.queryParams.subscribe(params => {
+      this.searchQuery = this.parseParameters(params);
+      this.storeFinderService.findStores(
+        this.searchQuery.queryText,
+        this.searchQuery.useMyLocation
+      );
+    });
+
+    this.isLoading$ = this.storeFinderService.getStoresLoading();
+    this.locations$ = this.storeFinderService
+      .getFindStoresEntities()
+      .pipe(tap(data => (this.geolocation = data.geolocation)));
   }
 
-  ngOnDestroy(): void {
-    if (this.ngUnsubscribe) {
-      this.ngUnsubscribe.unsubscribe();
+  ngOnDestroy() {
+    if (this.subscription) {
+      this.subscription.unsubscribe();
     }
   }
 
   viewPage(pageNumber: number) {
     this.searchConfig = { ...this.searchConfig, currentPage: pageNumber };
-    this.store.dispatch(
-      new fromStore.FindStores({
-        queryText: this.searchQuery.queryText,
-        longitudeLatitude: this.geolocation,
-        searchConfig: this.searchConfig
-      })
-    );
-  }
-
-  private initialize(params: Params) {
-    this.searchQuery = this.parseParameters(params);
-    this.storeFinderService.findStores(
+    this.storeFinderService.findStoresAction(
       this.searchQuery.queryText,
-      this.searchQuery.useMyLocation
+      this.geolocation,
+      this.searchConfig
     );
-
-    this.isLoading$ = this.store.pipe(select(fromStore.getStoresLoading));
-    this.locations$ = this.store.pipe(select(fromStore.getFindStoresEntities));
-    this.ngUnsubscribe = this.locations$
-      .pipe(map(data => data.geolocation))
-      .subscribe(geoData => (this.geolocation = geoData));
   }
 
-  private parseParameters(queryParams: { [key: string]: any }): SearchQuery {
-    let searchQuery: SearchQuery;
+  private parseParameters(queryParams: {
+    [key: string]: any;
+  }): StoreFinderSearchQuery {
+    let searchQuery: StoreFinderSearchQuery;
 
     if (queryParams.query) {
       searchQuery = { queryText: queryParams.query };
