@@ -1,5 +1,5 @@
 import { Input } from '@angular/core';
-import { StoreDataService } from '../../services/index';
+import { StoreDataService } from '@spartacus/core';
 
 export class AbstractStoreItemComponent {
   @Input()
