@@ -1,23 +1,16 @@
 import { ActivatedRoute } from '@angular/router';
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, async } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { StoreModule } from '@ngrx/store';
 
 import { StoreFinderGridComponent } from './store-finder-grid.component';
-import { StoreFinderService } from '../../services/store-finder.service';
 import { SpinnerModule } from '../../../ui/components/spinner/spinner.module';
 
-import * as fromReducers from '../../store';
+import { StoreFinderService, RoutingService } from '@spartacus/core';
 import { Pipe, PipeTransform, Component, Input } from '@angular/core';
-import { RoutingService } from '@spartacus/core';
+import { of, Observable } from 'rxjs';
 
 const countryIsoCode = 'CA';
 const regionIsoCode = 'CA-QC';
-
-class StoreFinderServiceMock {
-  viewAllStoresForCountry(_countryIso: string) {}
-  viewAllStoresForRegion(_countryIso: string, _regionIso: string) {}
-}
 
 @Component({
   selector: 'cx-store-finder-list-item',
@@ -38,6 +31,13 @@ const mockActivatedRoute = {
   }
 };
 
+const mockStoreFinderService = {
+  viewAllStoresForCountry: jasmine.createSpy(),
+  viewAllStoresForRegion: jasmine.createSpy(),
+  getStoresLoading: jasmine.createSpy(),
+  getFindStoresEntities: jasmine.createSpy().and.returnValue(of(Observable))
+};
+
 @Pipe({
   name: 'cxTranslateUrl'
 })
@@ -49,16 +49,38 @@ describe('StoreFinderGridComponent', () => {
   let component: StoreFinderGridComponent;
   let fixture: ComponentFixture<StoreFinderGridComponent>;
   let storeFinderService: StoreFinderService;
+  let route: ActivatedRoute;
+
   const mockRoutingService = {
     go: jasmine.createSpy('go')
   };
 
-  it('should create with country routing parameter', () => {
-    mockActivatedRoute.snapshot.params = { country: countryIsoCode };
-    configureTestBed();
-    spyOn(storeFinderService, 'viewAllStoresForCountry');
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [RouterTestingModule, SpinnerModule],
+      declarations: [
+        StoreFinderGridComponent,
+        MockStoreFinderListItemComponent,
+        MockTranslateUrlPipe
+      ],
+      providers: [
+        { provide: StoreFinderService, useValue: mockStoreFinderService },
+        { provide: ActivatedRoute, useValue: mockActivatedRoute },
+        { provide: RoutingService, useValue: mockRoutingService }
+      ]
+    }).compileComponents();
+  }));
 
-    createComponent();
+  beforeEach(() => {
+    fixture = TestBed.createComponent(StoreFinderGridComponent);
+    component = fixture.componentInstance;
+    route = TestBed.get(ActivatedRoute);
+    storeFinderService = TestBed.get(StoreFinderService);
+  });
+
+  it('should create with country routing parameter', () => {
+    route.snapshot.params = { country: countryIsoCode };
+    fixture.detectChanges();
 
     expect(component).toBeTruthy();
     expect(storeFinderService.viewAllStoresForCountry).toHaveBeenCalledWith(
@@ -67,26 +89,21 @@ describe('StoreFinderGridComponent', () => {
   });
 
   it('should create with country and region routing parameters', () => {
-    mockActivatedRoute.snapshot.params = {
+    route.snapshot.params = {
       country: countryIsoCode,
       region: regionIsoCode
     };
-    configureTestBed();
-    spyOn(storeFinderService, 'viewAllStoresForRegion');
-
-    createComponent();
+    fixture.detectChanges();
 
     expect(component).toBeTruthy();
   });
 
   it('should route when viewStore is called with region', () => {
-    mockActivatedRoute.snapshot.params = {
+    route.snapshot.params = {
       country: countryIsoCode,
       region: regionIsoCode
     };
-    configureTestBed();
-    spyOn(storeFinderService, 'viewAllStoresForRegion');
-    createComponent();
+    fixture.detectChanges();
 
     component.viewStore(location);
 
@@ -110,12 +127,10 @@ describe('StoreFinderGridComponent', () => {
   });
 
   it('should route when viewStore is called without region', () => {
-    mockActivatedRoute.snapshot.params = {
+    route.snapshot.params = {
       country: countryIsoCode
     };
-    configureTestBed();
-    spyOn(storeFinderService, 'viewAllStoresForCountry');
-    createComponent();
+    fixture.detectChanges();
 
     component.viewStore(location);
 
@@ -125,6 +140,7 @@ describe('StoreFinderGridComponent', () => {
         {
           name: 'storeDescription',
           params: {
+            region: '',
             country: countryIsoCode,
             store: location.name
           }
@@ -132,34 +148,4 @@ describe('StoreFinderGridComponent', () => {
       ]
     });
   });
-
-  function configureTestBed(): void {
-    const bed = TestBed.configureTestingModule({
-      imports: [
-        StoreModule.forRoot({}),
-        StoreModule.forFeature('stores', fromReducers.reducers),
-        RouterTestingModule,
-        SpinnerModule
-      ],
-      declarations: [
-        StoreFinderGridComponent,
-        MockStoreFinderListItemComponent,
-        MockTranslateUrlPipe
-      ],
-      providers: [
-        { provide: StoreFinderService, useClass: StoreFinderServiceMock },
-        { provide: ActivatedRoute, useValue: mockActivatedRoute },
-        { provide: RoutingService, useValue: mockRoutingService }
-      ]
-    });
-    bed.compileComponents();
-
-    storeFinderService = bed.get(StoreFinderService);
-  }
-
-  function createComponent() {
-    fixture = TestBed.createComponent(StoreFinderGridComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  }
 });
