@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { DebugElement, SimpleChange, ElementRef } from '@angular/core';
 
 import { StoreFinderMapComponent } from './store-finder-map.component';
-import { GoogleMapRendererService } from '../../services/google-map-renderer.service';
+import { GoogleMapRendererService } from '@spartacus/core';
 
 class MapRendererServiceMock {
   public renderMap(_mapElement: HTMLElement, _locations: any[]): void {}
