@@ -1,5 +1,4 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { combineReducers, Store, StoreModule } from '@ngrx/store';
 import { RouterTestingModule } from '@angular/router/testing';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
@@ -7,12 +6,9 @@ import { By } from '@angular/platform-browser';
 
 import { StoreFinderListComponent } from './store-finder-list.component';
 import { StoreFinderMapComponent } from '../../store-finder-map/store-finder-map.component';
-import { StoreDataService } from '../../../services';
-import { GoogleMapRendererService } from '../../../services/google-map-renderer.service';
 import { SpinnerModule } from '../../../../ui/components/spinner/spinner.module';
 
-import * as fromReducers from '../../../store';
-import * as fromServices from '../../../services';
+import { StoreDataService, GoogleMapRendererService } from '@spartacus/core';
 
 const location = {};
 const stores = [location];
@@ -36,25 +32,16 @@ class GoogleMapRendererServiceMock {
 describe('StoreFinderDisplayListComponent', () => {
   let component: StoreFinderListComponent;
   let fixture: ComponentFixture<StoreFinderListComponent>;
-  let store: Store<fromReducers.StoresState>;
   let storeMapComponent: StoreFinderMapComponent;
   let storeDataService: StoreDataService;
   let googleMapRendererService: GoogleMapRendererService;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [
-        RouterTestingModule,
-        HttpClientTestingModule,
-        SpinnerModule,
-        StoreModule.forRoot({
-          stores: combineReducers(fromReducers.reducers)
-        })
-      ],
+      imports: [RouterTestingModule, HttpClientTestingModule, SpinnerModule],
       schemas: [NO_ERRORS_SCHEMA],
       declarations: [StoreFinderListComponent, StoreFinderMapComponent],
       providers: [
-        ...fromServices.services,
         {
           provide: GoogleMapRendererService,
           useClass: GoogleMapRendererServiceMock
@@ -67,11 +54,9 @@ describe('StoreFinderDisplayListComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(StoreFinderListComponent);
     component = fixture.componentInstance;
-    store = TestBed.get(Store);
     storeDataService = TestBed.get(StoreDataService);
     googleMapRendererService = TestBed.get(GoogleMapRendererService);
 
-    spyOn(store, 'dispatch');
     spyOn(storeDataService, 'getStoreLatitude');
     spyOn(storeDataService, 'getStoreLongitude');
     spyOn(googleMapRendererService, 'centerMap');
