@@ -1,13 +1,12 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { Store, StoreModule } from '@ngrx/store';
 import { RouterTestingModule } from '@angular/router/testing';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 
-import { StoreFinderService } from '../../services';
 import { StoreFinderSearchResultComponent } from './store-finder-search-result.component';
 
-import * as fromReducers from '../../store';
+import { StoreFinderService } from '@spartacus/core';
+import { Observable, of } from 'rxjs';
 
 class ActivatedRouteMock {
   paramsSubscriptionHandler: Function;
@@ -19,30 +18,27 @@ class ActivatedRouteMock {
   };
 }
 
-class StoreFinderServiceMock {
-  findStores() {}
-}
+const mockStoreFinderService = {
+  findStores: jasmine.createSpy(),
+  getStoresLoading: jasmine.createSpy(),
+  getFindStoresEntities: jasmine.createSpy().and.returnValue(of(Observable)),
+  findStoresAction: jasmine.createSpy().and.returnValue(of({}))
+};
 
 describe('StoreFinderListComponent', () => {
   let component: StoreFinderSearchResultComponent;
   let fixture: ComponentFixture<StoreFinderSearchResultComponent>;
-  let store: Store<fromReducers.StoresState>;
   let storeFinderService: StoreFinderService;
   let activatedRoute: ActivatedRouteMock;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [
-        RouterTestingModule,
-        StoreModule.forRoot({}),
-        StoreModule.forFeature('stores', fromReducers.reducers)
-      ],
+      imports: [RouterTestingModule],
       schemas: [NO_ERRORS_SCHEMA],
       declarations: [StoreFinderSearchResultComponent],
       providers: [
-        { provide: StoreFinderService, useClass: StoreFinderServiceMock },
-        { provide: ActivatedRoute, useClass: ActivatedRouteMock },
-        Store
+        { provide: StoreFinderService, useValue: mockStoreFinderService },
+        { provide: ActivatedRoute, useClass: ActivatedRouteMock }
       ]
     }).compileComponents();
   }));
@@ -50,12 +46,8 @@ describe('StoreFinderListComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(StoreFinderSearchResultComponent);
     component = fixture.componentInstance;
-    store = TestBed.get(Store);
     storeFinderService = TestBed.get(StoreFinderService);
     activatedRoute = TestBed.get(ActivatedRoute);
-
-    spyOn(store, 'dispatch');
-    spyOn(storeFinderService, 'findStores');
 
     fixture.detectChanges();
   });
@@ -103,12 +95,10 @@ describe('StoreFinderListComponent', () => {
     component.viewPage(pageNumber);
 
     // then
-    expect(store.dispatch).toHaveBeenCalledWith(
-      new fromReducers.FindStores({
-        queryText: '',
-        longitudeLatitude: { longitude: 0, latitude: 0 },
-        searchConfig: { currentPage: pageNumber }
-      })
+    expect(storeFinderService.findStoresAction).toHaveBeenCalledWith(
+      '',
+      { longitude: 0, latitude: 0 },
+      { currentPage: pageNumber }
     );
   });
 });
