@@ -1,12 +1,10 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { StoreModule } from '@ngrx/store';
 
 import { SpinnerModule } from '../../../ui/components/spinner/spinner.module';
 import { StoreFinderStoresCountComponent } from './store-finder-stores-count.component';
-import { StoreFinderService } from '../../services';
+import { RoutingService, StoreFinderService } from '@spartacus/core';
 
-import * as fromReducers from '../../store';
 import { Pipe, PipeTransform } from '@angular/core';
 
 @Pipe({
@@ -16,20 +14,30 @@ class MockTranslateUrlPipe implements PipeTransform {
   transform() {}
 }
 
+const mockStoreFinderService = {
+  viewAllStores: jasmine.createSpy(),
+  getViewAllStoresEntities: jasmine.createSpy(),
+  getViewAllStoresLoading: jasmine.createSpy()
+};
+
 describe('StoreFinderStoresCountComponent', () => {
   let component: StoreFinderStoresCountComponent;
   let fixture: ComponentFixture<StoreFinderStoresCountComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [
-        SpinnerModule,
-        StoreModule.forRoot({}),
-        StoreModule.forFeature('stores', fromReducers.reducers),
-        RouterTestingModule
-      ],
+      imports: [SpinnerModule, RouterTestingModule],
       declarations: [StoreFinderStoresCountComponent, MockTranslateUrlPipe],
-      providers: [StoreFinderService]
+      providers: [
+        {
+          provide: RoutingService,
+          useValue: { go: jasmine.createSpy() }
+        },
+        {
+          provide: StoreFinderService,
+          useValue: mockStoreFinderService
+        }
+      ]
     }).compileComponents();
   }));
 
